+++ pkgs/timezone/lib/src/exceptions.dart
@@ -2,7 +2,7 @@
 // file for details. All rights reserved. Use of this source code is governed
 // by a BSD-style license that can be found in the LICENSE file.
 
-part of timezone;
+library timezone.src.exceptions;
 
 class TimeZoneInitException implements Exception {
   final String msg;