+++ pkgs/timezone/lib/src/tools.dart
@@ -2,7 +2,7 @@
 // file for details. All rights reserved. Use of this source code is governed
 // by a BSD-style license that can be found in the LICENSE file.
 
-library timezone.tools;
+library timezone.src.tools;
 
 import 'dart:collection';
 import 'package:tuple/tuple.dart';